Parallel stages, plugins and package installation...
The parallel stage was implemented with `multiprocessing`;

A plugin system using packages' metadata ('entry_point') was implemented. An
initial "core" plugin was created that provides an "sh" step/command.

To be able to run and test the plugin implementation, a `setup.py` was created
that exposes the internal plugin (to itself).

(Sorry for the single big commit)